Add file test_sectors_and_sector_concepts.omn #30
This file was created as prove of concept of my ideas regarding sectors and sector concepts. It is not meant to be included 1:1 into master or dev branches but some _elements_ might be included in the other branches.
The file was presented on the SzenDB project meeting on 2020-01-21. Further discussions on the ontology meeting in February 2020.