Fix SubclassOf for 'biofuel'
In pull request #952 the class `biofuel` got the new definition _A biofuel is a combustion fuel that has a biogenic origin._
The axiom `EquivalentTo 'combustion fuel' and ('has origin' some biogenic)` was added.
However, it was missed to change at the same time `SubclassOf fuel` to `SubclassOf combustion fuel`.

I think we can do here a quick fix without opening an issue for that.