Exclude competence question Q1 #872
With the implementation of issue #872 in pull request https://github.com/OpenEnergyPlatform/ontology/pull/1409/ the competency question 1 gets deprecated.
The question reads:
> Is  biofuel renewable fuel?

We now distinguish between `sustainable biofuel` and `non-sustainable biofuel` and only a `sustainable biofuel` is a `renewable fuel`.
We decided in OEO that we do not want to change the original competency question.
As workaround for a proper solution for the competency questions, Q1 is marked by a comment as deprecated.
Further, Q1 gets excluded from the script.